requirements.txt
whisper
pyannote.audio
transformers
pyyaml
tqdm
click
torch
pydub
librosa
soundfile